# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: DigiHostel API
  version: 0.1.0
  description: >
    Source contract for the DigiHostel API (ADR-007). Health check plus the
    first two real business workflows, Parent Leave Approval and Student Leave
    Request creation/viewing (docs/leave-approval-workflow.md).
servers:
  - url: /api/v1
paths:
...
  /students:
    get:
      operationId: searchStudents
      tags: [students]
      summary: >
        Staff-only: server-side student search (Phase 4, Prompt 8 — Student
        Operations Center)
      description: >
        Requires an AAL2 (MFA-verified) staff session with role
        reception_warden, hostel_admin, or super_admin — matching every
        other certified staff-only leave/student route's role set and
        AAL2/hostel-scope shape exactly. reception_warden and hostel_admin
        see only students in their own assigned hostel, resolved entirely
        server-side from the authenticated caller's own staff row — never a
        client-supplied filter; super_admin is unscoped. `q`, when
        supplied, matches a case-insensitive PREFIX against full_name OR
        roll_number — the only two fields this system's authoritative
        `students` schema indexes for search; no department/program/gender/
        phone/academic-year field exists anywhere in this schema; none is
        exposed. Deterministically ordered (the requested sort field plus
        the student's own id as a tie-breaker) and paginated server-side —
        never the caller's whole in-scope population downloaded and
        filtered client-side.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
            maxLength: 200
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
        - name: sortBy
          in: query
          required: false
          schema:
            type: string
            enum: [fullName, rollNumber]
            default: fullName
        - name: sortDir
          in: query
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        "200":
          description: The caller's operationally-visible, paginated student search result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentSearchResult"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
  /students/{rollNumber}:
    get:
      operationId: getStudentProfile
      tags: [students]
      summary: >
        Staff-only: read-only student profile (Phase 4, Prompt 8 — Student
        Operations Center)
      description: >
        Same AAL2/role/hostel-scope shape as GET /students. Returns only
        authoritative DigiHostel data — identity, hostel/room, linked
        parent/guardian contact (name, relationship, phone only — never
        internal ids or authentication identifiers), and the student's own
        most recent DigiHostel Hostel Leaving Request with its immutable
        approval-event timeline, reused unmutated from the certified Parent
        Approval / Exit Authorization workflow (never the separate KIIT SAP
        Holiday Request concept, which has no integration anywhere in this
        system). A roll number that does not exist, or that exists outside
        the caller's hostel scope, is deliberately indistinguishable (404,
        anti-enumeration) — identical shape to
        GET /leave-requests/{leaveRequestId}.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: rollNumber
          in: path
          required: true
          schema:
            type: string
            maxLength: 100
      responses:
        "200":
          description: The student's read-only operational profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Auth-issued JWT (ADR-014).
...
  schemas:
...
    LeaveRequestStatus:
      type: string
      description: >
        Full state vocabulary already defined by the schema
        (packages/db/src/schema/enums.ts, leave_request_status) — this
        endpoint set only ever produces approved/rejected transitions today;
        the escalation-notification states are included because they are
        part of the same enum, not because this task creates them.
      enum:
        - pending
        - father_notified
        - mother_notified
        - guardian_notified
        - approved
        - rejected
        - in_app_call
        - manual_verification
        - expired
...
    ParentRelationshipType:
      type: string
      description: Mirrors packages/db/src/schema/enums.ts's parent_relationship_type enum.
      enum: [father, mother, guardian]
...
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            currentStatus:
              $ref: "#/components/schemas/LeaveRequestStatus"
            requestId:
              type: string
              description: >-
                Server-generated per-request identifier (F-07 production
                observability). Present on unexpected (5xx) and
                framework-rejected (4xx) errors; hand this value to support
                to correlate a client-visible failure with server-side logs.
                Not present on typed domain errors that routes map to a
                response themselves (e.g. leave-request 404/409).
    StudentSearchResultItem:
      type: object
      description: >
        Deliberately narrower than the full `students` row — never
        `authUserId` (an internal identity-linkage field with no product
        purpose on a staff-facing read surface). No department/program/
        semester/gender/phone/photograph field is represented anywhere in
        this schema — none exists in the authoritative `students` table.
      required: [id, rollNumber, fullName, hostelId, hostelName, roomId, roomNumber]
      properties:
        id:
          type: string
          format: uuid
        rollNumber:
          type: string
        fullName:
          type: string
        hostelId:
          type: string
          format: uuid
          nullable: true
        hostelName:
          type: string
          nullable: true
        roomId:
          type: string
          format: uuid
          nullable: true
        roomNumber:
          type: string
          nullable: true
    StudentSearchResult:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/StudentSearchResultItem"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
    StudentGuardian:
      type: object
      description: >
        One linked parent/guardian, minimized to exactly what the Student
        Operations Center profile needs — never an internal id or
        authentication identifier.
      required: [fullName, relationshipType, phoneNumber]
      properties:
        fullName:
          type: string
        relationshipType:
          $ref: "#/components/schemas/ParentRelationshipType"
        phoneNumber:
          type: string
    StudentCurrentLeave:
      type: object
      description: >
        The student's own most recent DigiHostel Hostel Leaving Request —
        reused, unmutated, read-only data from the certified Parent
        Approval / Exit Authorization workflow. Never the separate KIIT SAP
        Holiday Request concept.
      required:
        - id
        - status
        - reason
        - startDate
        - endDate
        - createdAt
        - updatedAt
        - exitAuthorized
        - exitAuthorizedAt
        - returnRecorded
        - returnedAt
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/LeaveRequestStatus"
        reason:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        exitAuthorized:
          type: boolean
        exitAuthorizedAt:
          type: string
          format: date-time
          nullable: true
        returnRecorded:
          type: boolean
        returnedAt:
          type: string
          format: date-time
          nullable: true
    StudentTimelineEvent:
      type: object
      description: >
        One immutable leave_approval_events row for the student's current
        leave request — same actor-omission discipline as
        LeaveApprovalEvent: never reveals which specific parent/guardian/
        staff member acted.
      required: [id, eventType, response, occurredAt]
      properties:
        id:
          type: string
          format: uuid
        eventType:
          type: string
          enum: [notified, responded, escalated, expired, manual_override]
        response:
          type: string
          enum: [approved, rejected, no_response]
          nullable: true
        occurredAt:
          type: string
          format: date-time
    HostelPresence:
      type: string
      description: >
        Server-derived, never persisted. "outside_hostel" iff the student's
        own current leave request is genuinely exit-authorized and has no
        return recorded yet; "inside_hostel" otherwise (including "no leave
        request at all"). Never trust a client-supplied value for this —
        it is computed entirely server-side from leave_exit_authorizations
        and movements (Phase 4, Prompt 9 remediation).
      enum: [inside_hostel, outside_hostel]
    StudentProfile:
      type: object
      required:
        - id
        - rollNumber
        - fullName
        - hostelId
        - hostelName
        - roomId
        - roomNumber
        - guardians
        - currentLeave
        - timeline
        - hostelPresence
      properties:
        id:
          type: string
          format: uuid
        rollNumber:
          type: string
        fullName:
          type: string
        hostelId:
          type: string
          format: uuid
          nullable: true
        hostelName:
          type: string
          nullable: true
        roomId:
          type: string
          format: uuid
          nullable: true
        roomNumber:
          type: string
          nullable: true
        guardians:
          type: array
          items:
            $ref: "#/components/schemas/StudentGuardian"
        currentLeave:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/StudentCurrentLeave"
        timeline:
          type: array
          items:
            $ref: "#/components/schemas/StudentTimelineEvent"
        hostelPresence:
          $ref: "#/components/schemas/HostelPresence"
...
  responses:
    Unauthenticated:
      description: Missing, malformed, or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    Forbidden:
      description: Authenticated but not authorized for this action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    NotFound:
      description: >
        Leave request does not exist, or exists but the caller has no
        relationship to its student (deliberately indistinguishable —
        anti-enumeration).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
...
    ValidationError:
      description: Request body/params failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"